refactor: Move OpenLectureRaw to data module
The OpenLectureRaw data class and its associated RemoteOpenLectureDataSource interface have been moved from the `timetable` package to the `openlecture` package within the `data` module.

This change aims to better organize the data-related components by grouping OpenLecture specific items together.

The following files were affected:
- `RemoteTimetableDataSourceModule.kt` updated to reflect the new path.
- `RemoteOpenLectureDataSourceImpl.kt` updated to reflect the new path.
- `OpenLectureRepositoryImpl.kt` updated to reflect the new path for `RemoteOpenLectureDataSource`.
- `OpenLectureRaw.kt` deleted from its old location.
- `RemoteOpenLectureDataSource.kt` deleted from its old location.
