+++ composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/data/openlecture/repository/OpenLectureRepositoryImpl.kt
@@ -13,71 +13,73 @@ import kotlinx.coroutines.flow.map
 import kotlinx.datetime.LocalDateTime
 
 class OpenLectureRepositoryImpl(
-  private val remoteOpenLectureDataSource: RemoteOpenLectureDataSource,
-  private val localOpenLectureDataSource: LocalOpenLectureDataSource,
+    private val remoteOpenLectureDataSource: RemoteOpenLectureDataSource,
+    private val localOpenLectureDataSource: LocalOpenLectureDataSource,
 ) : OpenLectureRepository {
-  override suspend fun getOpenLectureList(
-    lectureOrProfessorName: String?,
-    major: String?,
-    grade: Int?,
-  ): List<OpenLecture> {
-    return localOpenLectureDataSource.getOpenLectureList(
-      lectureOrProfessorName = lectureOrProfessorName,
-      major = major,
-      grade = grade,
-    )
-  }
-
-  override suspend fun checkNeedUpdate(): Boolean {
-    val localVersion = localOpenLectureDataSource.getOpenLectureListVersion().firstOrNull() ?: return true
-    val remoteVersion = remoteOpenLectureDataSource.getOpenLectureListVersion()
-    return remoteVersion > localVersion
-  }
-
-  override suspend fun updateAllLectures() = coroutineScope {
-    val remoteOpenLectures = async {
-      remoteOpenLectureDataSource.getOpenLectureList().map {
-        OpenLecture(
-          id = it.number,
-          name = it.className,
-          type = it.classification,
-          major = it.major,
-          grade = it.grade,
-          professorName = it.professor,
-          originalCellList = TimetableUtil.parseTimeTableString(it.time),
+    override suspend fun getOpenLectureList(
+        lectureOrProfessorName: String?,
+        major: String?,
+        grade: Int?,
+    ): List<OpenLecture> {
+        return localOpenLectureDataSource.getOpenLectureList(
+            lectureOrProfessorName = lectureOrProfessorName,
+            major = major,
+            grade = grade,
         )
-      }
     }
 
-    val remoteOpenLectureVersion = async {
-      remoteOpenLectureDataSource.getOpenLectureListVersion()
+    override suspend fun checkNeedUpdate(): Boolean {
+        val localVersion =
+            localOpenLectureDataSource.getOpenLectureListVersion().firstOrNull() ?: return true
+        val remoteVersion = remoteOpenLectureDataSource.getOpenLectureListVersion()
+        return remoteVersion > localVersion
     }
 
-    localOpenLectureDataSource.updateAllLectures(remoteOpenLectures.await())
-    localOpenLectureDataSource.setOpenLectureListVersion(remoteOpenLectureVersion.await())
-  }
+    override suspend fun updateAllLectures() = coroutineScope {
+        val remoteOpenLectures = async {
+            remoteOpenLectureDataSource.getOpenLectureList().map {
+                OpenLecture(
+                    id = it.number,
+                    name = it.className,
+                    type = it.classification,
+                    major = it.major,
+                    grade = it.grade,
+                    professorName = it.professor,
+                    originalCellList = TimetableUtil.parseTimeTableString(it.time),
+                )
+            }
+        }
 
-  override suspend fun getLastUpdatedDate(): String? {
-    return try {
-      val version = localOpenLectureDataSource.getOpenLectureListVersion().firstOrNull()?.toString()
-      if (version == null || version.length != 12) return null
+        val remoteOpenLectureVersion = async {
+            remoteOpenLectureDataSource.getOpenLectureListVersion()
+        }
 
-      val year = version.substring(0, 4).toInt()
-      val month = version.substring(4, 6).toInt()
-      val day = version.substring(6, 8).toInt()
-      val hour = version.substring(8, 10).toInt()
-      val minute = version.substring(10, 12).toInt()
-
-      val dateTime = LocalDateTime(year, month, day, hour, minute)
-
-      "${dateTime.year}년 ${dateTime.monthNumber}월 ${dateTime.dayOfMonth}일 " +
-              "${dateTime.hour}시 ${dateTime.minute}분"
-    } catch (e: Exception) {
-      null
+        localOpenLectureDataSource.updateAllLectures(remoteOpenLectures.await())
+        localOpenLectureDataSource.setOpenLectureListVersion(remoteOpenLectureVersion.await())
     }
-  }
 
-  override suspend fun getOpenMajor(): List<String> {
-    return localOpenLectureDataSource.getOpenLectureList().map { list -> list.major }
-  }
+    override suspend fun getLastUpdatedDate(): String? {
+        return try {
+            val version =
+                localOpenLectureDataSource.getOpenLectureListVersion().firstOrNull()?.toString()
+            if (version == null || version.length != 12) return null
+
+            val year = version.substring(0, 4).toInt()
+            val month = version.substring(4, 6).toInt()
+            val day = version.substring(6, 8).toInt()
+            val hour = version.substring(8, 10).toInt()
+            val minute = version.substring(10, 12).toInt()
+
+            val dateTime = LocalDateTime(year, month, day, hour, minute)
+
+            "${dateTime.year}년 ${dateTime.monthNumber}월 ${dateTime.dayOfMonth}일 " +
+                    "${dateTime.hour}시 ${dateTime.minute}분"
+        } catch (e: Exception) {
+            null
+        }
+    }
+
+    override suspend fun getOpenMajor(): List<String> {
+        return localOpenLectureDataSource.getOpenLectureList().map { list -> list.major }
+    }
 }
