--- composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/data/timetable/model/OpenLectureRaw.kt
@@ -1,11 +1,0 @@
-package com.chukchukhaksa.mobile.data.timetable.model
-
-data class OpenLectureRaw(
-    val number: Long = 0, // 번호
-    val major: String = "", // 개설학과
-    val grade: Int = 1, // 개설학년
-    val className: String = "", // 과목명
-    val classification: String = "", // 이수 구분
-    val professor: String = "",
-    val time: String = "",
-)
