+++ composeApp/src/commonMain/kotlin/com/chukchukhaksa/mobile/remote/timetable/RemoteTimetableDataSourceModule.kt
@@ -1,6 +1,6 @@
 package com.chukchukhaksa.mobile.remote.timetable
 
-import com.chukchukhaksa.mobile.data.timetable.datasource.RemoteOpenLectureDataSource
+import com.chukchukhaksa.mobile.data.openlecture.datasource.RemoteOpenLectureDataSource
 import org.koin.dsl.module
 
 val remoteTimetableDataSourceModule = module {
